==> House-Prices-version-NicolasA/0-Introduction.ipynb <==
# Introduction au projet house-prices

https://www.kaggle.com/c/house-prices-advanced-regression-techniques/overview

Avec 79 variables explicatives décrivant (presque) tous les aspects des logements résidentiels à Ames, dans l'Iowa, ce concours  met au défi de prévoir le prix final de chaque logement.

from IPython.display import Image

Avant de commencer, j'ai lancé un opti-ML depuis l'interface de Big ML afin d'évaluer la performance du jeu de données à l état brut. 

Image("png/roptiml.png")

Ici le model ensemble(random decision forest) est le plus performant, cependant le model deepnet est présent en 2 émé position.

from biglearn import createNewPred as CP

CP.predOnProdSet()

Image("png/Kagg1.png")

Image("png/ldb.png")

Un premier test avait été effectué en tout début de formation sur un model simple, ici en passant par le modéle ensemble je gagne 800 places, sans avoir modifié le dataset. Ici on voit que l'opti-Ml permet de gagner beaucoup de temps et d'affiner le choix du modéle dés le départ.
