==> House-Prices-version-NicolasA/4-Dataset corrélation.ipynb <==
import pandas as pd
import numpy as np
from matplotlib import pyplot as plt
from sklearn.preprocessing import LabelEncoder
from sklearn.metrics import mean_squared_error
import seaborn as sns
import xgboost
from pandas import read_csv
import pandas
import numpy as np
import matplotlib.pyplot as plt
import csv

df_old = read_csv("Basetrain.csv")

df_old.shape

df_new = read_csv("new_Basetrain.csv")

df_new.shape

df_kagg = read_csv("new_kaggtest.csv")

df_kagg.shape

# Pour avoir sous la main toutes les colonnes afin de copier/coller plus facilement
for i in df_new :
    print(i)

# supréssion des lignes dont les corrélations sont hors PCA (voir PCA factor notebooks 2-)
x_list = [524,1299,636,186,183,170]
for i in x_list :
    df_new.drop(df_new[df_new.Id == i].index, inplace=True)

# Ajout de colonnes qui regroupent les metrics Garage 'qualitée' et les conditions du logement, sur le dataset train et test
# de kaggle
df_new["T_garage"] = df_new["GarageCond"]+df_new["GarageFinish"]+df_new["GarageQual"]+df_new["GarageCars"]
df_new["T_Overall"] = df_new["OverallCond"] + df_new["OverallQual"]
df_new["T_Ext"] = df_new["ExterCond"] + df_new["ExterQual"]

df_kagg["T_garage"] = df_kagg["GarageCond"]+df_kagg["GarageFinish"]+df_kagg["GarageQual"]+df_kagg["GarageCars"]
df_kagg["T_Overall"] = df_kagg["OverallCond"] + df_kagg["OverallQual"]
df_kagg["T_Ext"] = df_kagg["ExterCond"] + df_kagg["ExterQual"]

print(df_new["T_garage"],df_new["T_Overall"],df_kagg["T_garage"],df_kagg["T_Overall"])

df_new['SalePriceLog'] = np.log(df_new.target)

df_new.shape

df_kagg.shape

y_train = df_new.SalePriceLog.values.reshape((1453,1))
x_train = df_new.drop(['target','SalePriceLog'], axis=1)
x_test = df_kagg

xgb = xgboost.XGBRegressor(colsample_bytree=0.8, subsample=0.5,
                             learning_rate=0.05, max_depth=3, 
                             min_child_weight=1.8, n_estimators=2000,
                             reg_alpha=0.1, reg_lambda=0.3, gamma=0.01, 
                             silent=1, random_state =7, nthread = -1)


xgb.fit(x_train, y_train)
xgb_pred = xgb.predict(x_test)
y_test = xgb.predict(x_train)

RMSE = np.sqrt(mean_squared_error(y_train, y_test))
print(RMSE.round(4))

Image("png/4.png")

Image("png/5.png")

submission = pd.DataFrame({"Id": df_kagg['Id'],"SalePrice": np.exp(xgb_pred)})
submission.to_csv("new_kagg2.csv")

from IPython.display import Image
Image("png/Kagg3.png")

Image("png/idb3.png")

df_new.to_csv("new_Basetrain.csv")
df_kagg.to_csv("new_kaggtest.csv")

# Persévérance

df_new = read_csv("new_Basetrain.csv")
df_kagg = read_csv("new_kaggtest.csv")

df_new["T_heat"] = df_kagg["Heating"]+df_kagg["HeatingQC"]
df_new["T_Kittchen"] = df_new["KitchenAbvGr"] + df_new["KitchenQual"]
df_new["T_Condition"] = df_new["Condition1"] + df_new["Condition2"]
df_new["T_Ext"] = df_new["ExterCond"] + df_new["ExterQual"]
df_new["T_Land"] = df_new["LandContour"] + df_new["LandSlope"]
df_new["T_Roof"] = df_new["RoofMatl"] + df_new["RoofStyle"]
df_new["T_Fire"] = df_new["FireplaceQu"] + df_new["Fireplaces"]
df_new["T_Bsmt"] = df_new["BsmtCond"] + df_new["BsmtExposure"] + df_new["BsmtQual"]
df_new["T_Yearsfake"] = df_new["YearRemodAdd"]-df_new["YearBuilt"]

df_kagg["T_heat"] = df_kagg["Heating"]+df_kagg["HeatingQC"]
df_kagg["T_Kittchen"] = df_kagg["KitchenAbvGr"] + df_kagg["KitchenQual"]
df_kagg["T_Condition"] = df_kagg["Condition1"] + df_kagg["Condition2"]
df_kagg["T_Ext"] = df_kagg["ExterCond"] + df_kagg["ExterQual"]
df_kagg["T_Land"] = df_kagg["LandContour"] + df_kagg["LandSlope"]
df_kagg["T_Roof"] = df_kagg["RoofMatl"] + df_kagg["RoofStyle"]
df_kagg["T_Fire"] = df_kagg["FireplaceQu"] + df_kagg["Fireplaces"]
df_kagg["T_Bsmt"] = df_kagg["BsmtCond"] + df_kagg["BsmtExposure"] + df_kagg["BsmtQual"]
df_kagg["T_Yearsfake"] = df_kagg["YearRemodAdd"] - df_kagg["YearBuilt"]

df_new.drop(columns="PoolQC")
df_new.drop(columns="MiscFeature")
df_kagg.drop(columns="PoolQC")
df_kagg.drop(columns="MiscFeature")

df_new['SalePriceLog'] = np.log(df_new.target)

# Impact sur la target des nouvelles features

Image("png/6.png")

from IPython.display import Image
Image("png/7.png")

y_train = df_new.SalePriceLog.values.reshape((1453,1))
x_train = df_new.drop(['target','SalePriceLog'], axis=1)
x_test = df_kagg

xgb = xgboost.XGBRegressor(colsample_bytree=0.8, subsample=0.5,
                             learning_rate=0.05, max_depth=3, 
                             min_child_weight=1.8, n_estimators=2000,
                             reg_alpha=0.1, reg_lambda=0.3, gamma=0.01, 
                             silent=1, random_state =7, nthread = -1)


xgb.fit(x_train, y_train)
xgb_pred = xgb.predict(x_test)
y_test = xgb.predict(x_train)

RMSE = np.sqrt(mean_squared_error(y_train, y_test))
print(RMSE.round(4))

submission = pd.DataFrame({"Id": df_kagg['Id'],"SalePrice": np.exp(xgb_pred)})
submission.to_csv("new_kagg4.csv")

Cette opération de 'Persévérance' ne fait pas progresser le résultat Kaggle, elle fait un score inférieur, le dataset restera donc comme il était avant l étape perseverance. Il va falloir chercher d'autres piste.
