==> House-Prices-version-NicolasA/5-Entrainement 2.ipynb <==
import pandas as pd
import numpy as np
from matplotlib import pyplot as plt
from sklearn.preprocessing import LabelEncoder
from sklearn.metrics import mean_squared_error
import seaborn as sns
import xgboost
from pandas import read_csv
import pandas
import numpy as np
import matplotlib.pyplot as plt
import csv

df_new = read_csv("new_Basetrain.csv")
df_new.head(5)

df_kagg = read_csv("new_kaggtest.csv")
df_kagg.shape

df_new.info()

k = 10 #number of variables for heatmap
cols = corrmat.nlargest(k,'target')['target'].index
cm = np.corrcoef(df_new[cols].values.T)
sns.set(font_scale=1)
hm = sns.heatmap(cm, cbar=True, linewidths=.5, annot=True, square=True, fmt='.2f', annot_kws={'size': 10}, yticklabels=cols.values, xticklabels=cols.values)
plt.show()

from pandas import read_csv
from alys import regression as alysCL
from alys import  save as SV
from graph import rViz 
import pandas
import numpy as np
import matplotlib.pyplot as plt
import sys
import os
import csv
from biglearn import createNewPred as CP
from biglearn import analyserML as ML

Retour sous BigMl avec le nouveau dataset, je vais evaluer la moyenne de la marge par erreur pour la comparer à la situation initiale

df_new = read_csv("new_Basetrain.csv")
df_new = df_new.drop(columns = "SalePriceLog")
df_new.to_csv("new_Basetrain2.csv")

CP.newDataSet()

df = read_csv('Pred_Files/ML_pred4.csv')
alysCL.regr(df)

alysCL.posneg(df)

df.sort_values("errorabs", axis = 0, ascending = False)

SV.extraction(df)

alysCL.seuil_vs_error(df)

alysCL.seuil_vs_error(df)

count = 0
for i in df["Id"] :
    count +=1

s = pandas.Series(range(1,(count +1)))
tablex=s    
    
tabley=df["errorabs"]

for i in tablex :
    df["seuil_applique"] = 0.1
    
tabley2 =df["seuil_applique"]
rViz.graph_double(tablex,tabley,tabley2)

La moyenne de la marge par erreur a diminuée, grâce aux améliorations appliqué sur le dataframe
