==> House-Prices-version-NicolasA/9-House-Prices-Keras.ipynb <==
%autosave 0
import numpy as np 
import pandas as pd 
from pandas import read_csv, DataFrame
from math import sqrt

from sklearn.preprocessing import StandardScaler
from sklearn.model_selection import train_test_split
from sklearn.model_selection import GridSearchCV
from sklearn.metrics import mean_squared_error as MSE
from sklearn.model_selection import StratifiedKFold

from sklearn.preprocessing import PolynomialFeatures 
from sklearn.pipeline import Pipeline

import keras
from keras.models import Sequential
from keras import Sequential
from keras.layers import Dense, Dropout, Input, Activation
from keras import metrics
from keras import optimizers
from keras import regularizers

from keras.wrappers.scikit_learn import KerasRegressor

import seaborn as sns
import matplotlib.pyplot as plt

# ignore Warning
import warnings
warnings.filterwarnings("ignore") 



#### Mettre la variable sur True si recherche des paramétres opti, mais l'opération dure environ 45 minutes
lookForOptimiz = True

df_train = read_csv('new_Basetrain.csv')
df_Kagg = read_csv('new_kaggtest.csv')

submission = pd.DataFrame()
submission["Id"]= df_Kagg["Id"]

# Preparation du split

X_train = df_train

V_train = df_train['target'].values
F_train = df_train.drop(['target','SalePriceLog'], axis=1)

F_train

V_train

seed = 42
np.random.seed(seed)

X_train, X_val, y_train, y_val = train_test_split(F_train, V_train, test_size=0.3, random_state=seed)

scale = StandardScaler()
X_train = scale.fit_transform(X_train)
X_val = scale.transform(X_val)
df_Kagg = scale.transform(df_Kagg)

y_scale = StandardScaler()
y_scale.fit(y_train.reshape(-1, 1))
y_train = y_scale.transform(y_train.reshape(-1, 1))
y_val_resc = y_scale.transform(y_val.reshape(-1, 1))

df_Kagg

# Recherche des paramétres opti du réseau

sgd = optimizers.SGD(lr=0.01, clipvalue=0.5)

def create_model():
    model = Sequential()
    model.add(Dense(200, input_dim=X_train.shape[1], kernel_regularizer=regularizers.l1(0.01)))
    model.add(Activation('relu')) 
#     model.add(Dense(100, kernel_initializer='normal'))
#     model.add(Activation('relu'))
    model.add(Dense(1, kernel_initializer='normal',kernel_regularizer=regularizers.l1(0.01)))
    model.compile(loss='mean_squared_error', optimizer=sgd)
    return model

def BestParams():
    estimators = []
    #estimators.append(('pf', PolynomialFeatures(interaction_only=True,include_bias=False)))
    estimators.append(('ss', StandardScaler()))
    estimators.append(('nn', KerasRegressor(build_fn=create_model, epochs=10, batch_size=5, verbose=0)))
    nn_pipe = Pipeline(estimators)

    nn_param_grid = {'nn__epochs': [1000,1500,2000],'nn__batch_size':[5,100,200]}

    gs_nn = GridSearchCV(nn_pipe, nn_param_grid,verbose=0, cv=3)
    print('Ok')
    gs_nn.fit(X_train, y_train)
    return gs_nn.best_params_

X_train.shape, y_train.shape

if lookForOptimiz : 
    best = BestParams()
    print(best)

# Creation du modèle

model = create_model()

m_fti = model.fit(X_train, y_train, epochs=2000, batch_size=200)

DataFrame(m_fti.history).plot()
plt.grid = True
plt.show

X_val.shape

pred = model.predict(X_val)
pred = y_scale.inverse_transform(pred)

RMSE = np.sqrt(MSE(pred,y_val))
print(RMSE.round(4))

pred_kagg = model.predict(df_Kagg)

pred_kagg.shape

pred_kagg = y_scale.inverse_transform(pred_kagg)

pred_kagg


submission["SalePrice"]= pred_kagg
submission.to_csv("new_kagg10.csv", index= False)

from IPython.display import Image
Image('png/Neuro.png')

------ Résultat Kaggle 0.1309 ---------

On est encore dans des valeurs de paramétre extréme, le modèle peut être encore optimisé
Best param = batch_size': 200, 'epochs': 2000

------ Résultat Kaggle 0.1347 ---------

standardscaler aprés split et inverse_transform
regularization l1 
Best param = batch_size': 100, 'epochs': 1000

------ Résultat Kaggle 0.16 ---------
Best param = batch_size': 5, 'epochs': 500

------ Précedent Kaggle 0.22 ---------
